pyproject.toml
[tool.poetry]
name = "pseud"
version = "2.0.0dev"
description = "Bidirectionnal RPC Api on top of pyzmq"
authors = ["Nicolas Delaby <ticosax@free.fr>"]
license = "Apache Software License"
repository = "https://github.com/ticosax/pseud"
readme = "README.rst"

[tool.poetry.group.main.dependencies]
python = ">=3.9"
pyzmq = "*"
msgpack-python = "*"
"zope.component" = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-pep8 = "*"
tox = "*"
pytest-cov = "*"
pdbpp = "*"
black = "*"
tox-pyenv = "*"
tox-poetry = "*"
ruff = "*"

[tool.black]
line-length = 89
skip-string-normalization = true
exclude = "docs/source/conf.py|.tox/|.git/|dist/.pytest_cache/|setup.py"

[tool.isort]
profile = "black"
line_length = 89

[tool.ruff]
line-length = 89
target-version = "py39"
select = ["E", "F", "B", "UP", "I"]

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
